move use_tools to root
